fix structure files
