+++ database/seeders/20200822202809-category-seeder.js
@@ -1,9 +1,10 @@
 'use strict';
-const Factor = require("../class_seeders/index")
+// const FactoriesFakesData = require("../FactoriesFakesDatas/index")
+const CategoryFactory = require("../factories/CategoryFactory")
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-   return queryInterface.bulkInsert('categories',Factor.getCategory());
+   return queryInterface.bulkInsert('categories',CategoryFactory(6));
   },
   down: async (queryInterface, Sequelize) => {
    return queryInterface.bulkDelete('categories', null, {});
