+++ database/seeders/20200827012736-course-seeder.js
@@ -1,9 +1,9 @@
 'use strict';
-const Factor = require("../class_seeders/index")
+const CoursesFactory = require("../factories/CoursesFactory")
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert('courses',Factor.getCourse());
+    return queryInterface.bulkInsert('courses',CoursesFactory(3));
   },
 
   down: async (queryInterface, Sequelize) => {
