+++ database/seeders/20200827024516-modules-seeder.js
@@ -1,9 +1,9 @@
 'use strict';
-const Factor = require("../class_seeders/index")
+const ModulesFactory = require("../factories/ModulesFactory")
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert('modules',Factor.getModule());
+    return queryInterface.bulkInsert('modules',ModulesFactory(3));
   },
 
   down: async (queryInterface, Sequelize) => {
