+++ database/seeders/20200827035144-studen-seeder.js
@@ -1,9 +1,9 @@
 'use strict';
-const Factor = require("../class_seeders/index")
+const StudentsFactory = require("../factories/StudentsFactory")
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert('students', Factor.getStudent());
+    return queryInterface.bulkInsert('students',StudentsFactory(3));
   },
 
   down: async (queryInterface, Sequelize) => {
